Make heartbeat (frame_max, channel_max) negotiation consistent with reference client
Java client and some others (Bunny, .NET) use a slightly different
negotiation logic. See #874 comments for details.

With this change, in some cases when previously heartbeats were disabled,
they will be enabled now. Arguably this is an improvement because disabling
heartbeats is rarely a good idea (unless all machines run with sensible
TCP keepalive settings, which is a rare thing to come by).

Closes #874.